+++ server/t_server-internalerr_test.go
@@ -84,10 +84,10 @@ func internalerrclient(sn string, t *testing.T) {
 	if len(resp) != 1 && resp[0] != 255 {
 		t.Errorf("len resp should 1 & resp[0] should be 255, but got len %d and '%v'", len(resp), string(resp))
 	}
-	if err.(*p.Status).Code != p.Errs["reqerr"].Code {
-		t.Errorf("err.Code should be %d but is %v", p.Errs["reqerr"].Code, err.(*p.Status).Code)
+	if err.(*p.Status).Code != p.Stats["reqerr"].Code {
+		t.Errorf("err.Code should be %d but is %v", p.Stats["reqerr"].Code, err.(*p.Status).Code)
 	}
-	if err.(*p.Status).Txt != p.Errs["reqerr"].Txt {
-		t.Errorf("err.Txt should be %s but is %v", p.Errs["reqerr"].Txt, err.(*p.Status).Txt)
+	if err.(*p.Status).Txt != p.Stats["reqerr"].Txt {
+		t.Errorf("err.Txt should be %s but is %v", p.Stats["reqerr"].Txt, err.(*p.Status).Txt)
 	}
 }